// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export interface Scalars {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSONObject: { input: any; output: any; }
  Time: { input: any; output: any; }
}

export interface ApiError extends Error {
  __typename?: 'ApiError';
  args?: Maybe<Scalars['JSONObject']['output']>;
  code?: Maybe<Scalars['Int']['output']>;
  data?: Maybe<Scalars['JSONObject']['output']>;
  message?: Maybe<Scalars['String']['output']>;
  path?: Maybe<Scalars['String']['output']>;
}
// ...
export interface Error {
  code?: Maybe<Scalars['Int']['output']>;
  data?: Maybe<Scalars['JSONObject']['output']>;
  message?: Maybe<Scalars['String']['output']>;
}
// ...
export interface ImportUserRightResult {
  __typename?: 'ImportUserRightResult';
  errors?: Maybe<Array<ApiError>>;
  id?: Maybe<Scalars['String']['output']>;
  message?: Maybe<Scalars['String']['output']>;
  name?: Maybe<Scalars['String']['output']>;
}
// ...
export interface UserRole {
  __typename?: 'UserRole';
  name?: Maybe<Scalars['String']['output']>;
  readonly?: Maybe<Scalars['Int']['output']>;
  roleid: Scalars['Int']['output'];
  rules?: Maybe<UserRoleRules>;
  type?: Maybe<Scalars['Int']['output']>;
}

export interface UserRoleInput {
  name?: InputMaybe<Scalars['String']['input']>;
  readonly?: InputMaybe<Scalars['Int']['input']>;
  rules?: InputMaybe<UserRoleRulesInput>;
  type?: InputMaybe<Scalars['Int']['input']>;
}

export interface UserRoleModule {
  __typename?: 'UserRoleModule';
  id?: Maybe<Scalars['String']['output']>;
  moduleid?: Maybe<Scalars['String']['output']>;
  relative_path?: Maybe<Scalars['String']['output']>;
  status?: Maybe<Scalars['Int']['output']>;
}

export interface UserRoleModuleInput {
  id?: InputMaybe<Scalars['String']['input']>;
  moduleid?: InputMaybe<Scalars['String']['input']>;
  status?: InputMaybe<Scalars['Int']['input']>;
}

export interface UserRoleRule {
  __typename?: 'UserRoleRule';
  name?: Maybe<Scalars['String']['output']>;
  status?: Maybe<Scalars['Int']['output']>;
}

export interface UserRoleRuleInput {
  name?: InputMaybe<Scalars['String']['input']>;
  status?: InputMaybe<Scalars['Int']['input']>;
}

export interface UserRoleRules {
  __typename?: 'UserRoleRules';
  actions?: Maybe<Array<UserRoleRule>>;
  actions_default_access?: Maybe<Scalars['Int']['output']>;
  api?: Maybe<Array<Scalars['String']['output']>>;
  api_access?: Maybe<Scalars['Int']['output']>;
  api_mode?: Maybe<Scalars['Int']['output']>;
  modules?: Maybe<Array<UserRoleModule>>;
  modules_default_access?: Maybe<Scalars['Int']['output']>;
  ui?: Maybe<Array<UserRoleRule>>;
  ui_default_access?: Maybe<Scalars['Int']['output']>;
}

export interface UserRoleRulesInput {
  actions?: InputMaybe<Array<UserRoleRuleInput>>;
  actions_default_access?: InputMaybe<Scalars['Int']['input']>;
  api?: InputMaybe<Array<Scalars['String']['input']>>;
  api_access?: InputMaybe<Scalars['Int']['input']>;
  api_mode?: InputMaybe<Scalars['Int']['input']>;
  modules?: InputMaybe<Array<UserRoleModuleInput>>;
  modules_default_access?: InputMaybe<Scalars['Int']['input']>;
  ui?: InputMaybe<Array<UserRoleRuleInput>>;
  ui_default_access?: InputMaybe<Scalars['Int']['input']>;
}